Move class `gc_pointer` into class `collector`
